Connect new model to application
