+++ src/main/java/tddtrainer/gui/catalog/ExerciseSelectorController.java
@@ -32,8 +32,8 @@ import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.web.WebView;
-import tddtrainer.catalog.CatalogDataSource;
 import tddtrainer.catalog.Exercise;
+import tddtrainer.catalog.KatalogDataSource;
 import tddtrainer.events.ExerciseEvent;
 import tddtrainer.events.Views;
 import tddtrainer.events.gui.ShowSelectDialogRequest;
@@ -53,10 +53,10 @@ public class ExerciseSelectorController extends BorderPane {
     private EventBus bus;
 
     Logger logger = LoggerFactory.getLogger(ExerciseSelectorController.class);
-    private final CatalogDataSource datasource;
+    private final KatalogDataSource datasource;
 
     @Inject
-    public ExerciseSelectorController(FXMLLoader loader, EventBus bus, CatalogDataSource datasource) {
+    public ExerciseSelectorController(FXMLLoader loader, EventBus bus, KatalogDataSource datasource) {
         this.bus = bus;
         this.datasource = datasource;
         this.bus.register(this);
@@ -94,9 +94,10 @@ public class ExerciseSelectorController extends BorderPane {
 
         List<Exercise> catalog;
         try {
-            catalog = datasource.loadCatalog();
-        } catch (JsonSyntaxException | UnirestException e) {
-            showFailedToLoadCatalogAlert(e.getClass().getSimpleName(), e.getMessage());
+            catalog = datasource.loadKatalog();
+        } catch (RuntimeException e) {
+            Throwable cause = e.getCause() != null ? e.getCause() : e;
+            showFailedToLoadCatalogAlert(cause.getClass().getSimpleName(), cause.getMessage());
             catalog = new ArrayList<>();
         }
         exerciseList.setItems(FXCollections.observableArrayList(catalog));
